Modifications in the structure of the libraries that compose the toolbox
