+++ clustintime/Visualization.py
@@ -10,7 +10,7 @@ Created on Mon Sep 27 09:39:05 2021
 import numpy as np
 import matplotlib.pyplot as plt # For graphs
 import pandas as pd
-from scipy.signal import find_peaks
+
 import networkx as nx # creation, manipulation and study of the structure, dynamics and functions of complex networks
 from IPython.display import HTML, display
 
@@ -49,33 +49,14 @@ def plot_labels(labels, title, multi_tap = [], single_tap = [], TR = 0.5):
         
 
 
-def RSS_peaks(data, near):
-    RSS_1 = [np.sum(np.square(i))**0.5 for i in data]
-    peaks = find_peaks(RSS_1)[0]
-    new_peaks = [0]*2*near*len(peaks)
 
-    for i in range(len(peaks)):
-        if peaks[i] == 0:
-            new_peaks[0:2*near]=range(2*near)
-        elif peaks[i] == len(RSS_1):
-             new_peaks[(len(RSS_1)-near):(len(RSS_1)+near)] = range((len(RSS_1)-near),(len(RSS_1)+near))
-        else:
-            new_peaks[(peaks[i]-near):(peaks[i]+near)] = range((peaks[i]-near),(peaks[i]+near))
-    new_peaks = np.array(new_peaks)
-    new_peaks = new_peaks[new_peaks != 0]
-    new_peaks = np.array(list(set(new_peaks)))   
-    plt.plot(RSS_1)
-    plt.plot(np.array(RSS_1)[new_peaks])
-    plt.title('RSS of original data vs filtered RSS')
-    plt.legend(['Original RSS'], 'Filtered RSS')
-    
-    return new_peaks
 
 def show_table(labels):
     Cluster,Count = np.unique(labels, return_counts = True)
     array= [Cluster, Count, Count/len(labels)]
     display_table(np.transpose([['Cluster'], ['Count'], ['Percentage']]))
     display_table(np.transpose(array))
+    
     
 def plot_two_maps(map_1, map_2,title_1, title_2 ,single_tap = [], multi_tap = [], contrast = 1 ):
     fig = plt.figure(figsize=(32, 9))
